+++ geograph/__init__.py
@@ -1,0 +1,3 @@
+# pylint: disable=missing-module-docstring
+from geograph._version import __version__
+from geograph.geograph import ComponentGeoGraph, GeoGraph, HabitatGeoGraph
