+++ geograph/constants.py
@@ -12,13 +12,6 @@ constants_path = pathlib.Path(os.path.realpath(__file__))
 SRC_PATH = pathlib.Path(os.path.dirname(constants_path))
 PROJECT_PATH = pathlib.Path(os.path.dirname(SRC_PATH))
 
-
-# Data directory on GWS
-DATA_DIR = PROJECT_PATH / "data"
-# Polygon data of Chernobyl Exclusion Zone (CEZ)
-ROIS = DATA_DIR / "chernobyl" / "chernobyl_rois.geojson"
-
-
 # Coordinate reference systems (crs)
 WGS84 = "EPSG:4326"  # WGS84 standard crs (latitude, longitude)
 UTM35N = "EPSG:32635"  # https://epsg.io/32635 - preferred crs for chernobyl region
