+++ geograph/geotimeline.py
@@ -9,8 +9,8 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from src.binary_graph_operations import NodeMap, identify_graphs
-from src.geograph import GeoGraph
+from geograph import GeoGraph
+from geograph.binary_graph_operations import NodeMap, identify_graphs
 
 # type alias
 TimeStamp = Union[int, datetime.datetime]
