+++ geograph/tests/testdata/adjacent/full.gpkg
