+++ geograph/tests/testdata/adjacent/lower_left.gpkg
