+++ geograph/tests/testdata/adjacent/lower_right.gpkg
