+++ geograph/tests/testdata/adjacent/upper_left.gpkg
