+++ geograph/tests/testdata/adjacent/upper_right.gpkg
