+++ geograph/tests/testdata/overlapping/full.gpkg
