+++ geograph/tests/testdata/overlapping/lower_left.gpkg
