+++ geograph/tests/testdata/overlapping/lower_right.gpkg
