+++ geograph/tests/testdata/overlapping/upper_right.gpkg
