+++ geograph/utils/geopandas_utils.py
@@ -3,13 +3,12 @@ from typing import Dict, List
 
 import geopandas as gpd
 import networkx as nx
-from shapely.geometry import MultiPolygon
-
-from src.utils.polygon_utils import (
+from geograph.utils.polygon_utils import (
     connect_with_interior_bulk,
     connect_with_interior_or_edge_bulk,
     connect_with_interior_or_edge_or_corner_bulk,
 )
+from shapely.geometry import MultiPolygon
 
 # For switching identifiction mode in `identify_node`
 _BULK_SPATIAL_IDENTIFICATION_FUNCTION = {
