+++ geograph/visualisation/geoviewer.py
@@ -3,19 +3,19 @@
 from __future__ import annotations
 
 import logging
+import threading
+import time
 from typing import TYPE_CHECKING, List, Optional, Union
 
 import folium
+import geograph
 import ipyleaflet
 import ipywidgets as widgets
 import pandas as pd
 import traitlets
-import time
-import threading
-
-from src import geograph, metrics
-from src.constants import CHERNOBYL_COORDS_WGS84, WGS84
-from src.visualisation import (
+from geograph import metrics
+from geograph.constants import CHERNOBYL_COORDS_WGS84, WGS84
+from geograph.visualisation import (
     control_widgets,
     folium_utils,
     graph_utils,
