--- src/__init__.py
@@ -1,2 +1,0 @@
-# pylint: disable=missing-module-docstring
-from src._version import __version__
