--- src/_version.py
@@ -1,2 +1,0 @@
-"""Module to store the __version__ string."""
-__version__ = "0.0.1"
